+++ crates/ruma-common/CHANGELOG.md
@@ -19,6 +19,7 @@ Improvements:
 - Add `InitialStateEvent::{new, to_raw, to_raw_any}`
 - Add a convenience method to construct `RoomEncryptionEventContent` with the recommended defaults.
 - `PushCondition::EventMatch` and `FlattenedJson` now use escaped dotted paths, according to MSC3873
+- Add `FullStateEventContent::redact`
 
 # 0.11.3